Move addUserPresence to database
