+++ app/src/main/java/com/github/ybecker/epforuml/database/MockDatabase.kt
@@ -222,5 +222,7 @@ class MockDatabase : Database() {
         }
     }
 
-
+    override fun setUserPresence() {
+        TODO("Not yet implemented")
+    }
 }
